refactor: simplify verticals.auth imports, move date formatting away
