+++ ggshield/core/text_utils.py
@@ -1,4 +1,5 @@
 import sys
+from datetime import datetime
 from enum import Enum, auto
 from typing import Any, Dict, List, NamedTuple, Optional, Union
 
@@ -264,3 +265,13 @@ def file_diff_info(
         format_text(filename, STYLE["filename"]),
         ", ".join(incidents_count),
     )
+
+
+def get_pretty_date(dt: datetime) -> str:
+    """
+    convert the given datetime to the format September 1st 2022
+    """
+    month_suffix = {1: "st", 2: "nd", 3: "rd"}.get(
+        dt.day if dt.day < 20 else dt.day % 10, "th"
+    )
+    return dt.strftime(f"%B {dt.day}{month_suffix} %Y")
