+++ ggshield/verticals/auth/__init__.py
@@ -1,0 +1,7 @@
+from .oauth import OAuthClient, OAuthError
+
+
+__all__ = [
+    "OAuthClient",
+    "OAuthError",
+]
